modified:   index.html 	modified:   package.json 	modified:   public/app.js 	new file:   src/javascripts/visitor-counter.js
