+++ index.html
@@ -6,6 +6,7 @@
     <meta http-equiv="X-UA-Compatible" content="ie=edge" />
     <title>Alfasarana | Beranda</title>
     <script defer src="./public/app.js"></script>
+    <script defer type="module" src="./src/javascripts/visitor-counter.js"></script>
     <script defer src="./node_modules/flowbite/dist/flowbite.min.js"></script>
     <link rel="stylesheet" href="./public/output-style.css" />
   </head>
