+++ package.json
@@ -1,7 +1,6 @@
 {
   "name": "static-gallery-app",
   "version": "1.0.0",
-  "description": "",
   "main": "app.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
@@ -17,5 +16,6 @@
   },
   "dependencies": {
     "flowbite": "^2.3.0"
-  }
+  },
+  "description": ""
 }
